requirements.txt
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-admin-generator==2.6.0
django-import-export==3.3.7
django-utils2==3.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
fastcore==1.7.4
fastlite==0.0.10
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
MarkupPy==1.14
nodeenv==1.9.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pyright==1.1.378
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.5.1
python-multipart==0.0.9
python-utils==3.8.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlite-minutils==3.37.0.post3
sqlparse==0.5.1
starlette==0.38.4
tablib==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
xlrd==2.0.1
xlwt==1.3.0
